refactor : 정렬된 stations를 만드는 로직을 도메인으로 분리
