+++ src/main/java/wooteco/subway/docs/기능목록.md
@@ -114,10 +114,11 @@
 
 - [ ] 중복 이름,id 처리와 동일 구간 삽입에 대한 예외 분리하기
 - [x] 전체 id를 얻기위해 상행id, 하행 id를 두번 걸쳐서 db로부터 긁어오던것을 한번에 긁어오도록 수정하기
-- [ ] 레포지토리가 들고 있는 많은 로직을 도메인으로 이동시키기
 - [x] 메소드명이 좀 더 명확한 이름을 가지게 수정하기
 - [x] 상행선을 찾는 로직에서, 상행성을 찾지 못하는 경우에 대한 처리를 예외를 발생하도록 수정
+- [ ] 레포지토리가 들고 있는 많은 로직을 도메인으로 이동시키기
 - [ ] 서비스에 있는 로직을 도메인으로 이동하기
+    - [x] 정렬된 stations를 만드는 로직을 도메인으로 분리
