+++ src/main/java/wooteco/subway/line/service/SectionService.java
@@ -22,11 +22,11 @@ public class SectionService {
         this.sectionRepository = sectionRepository;
     }
 
-    @Transactional
     public void lineCreateAdd(final Long lineId, final SectionRequest sectionRequest) {
         validateLineId(lineId);
         validateStations(sectionRequest);
-        sectionRepository.save(lineId, sectionRequest.getUpStationId(), sectionRequest.getDownStationId(), sectionRequest.getDistance());
+        Section section = new Section(sectionRequest.getUpStationId(), sectionRequest.getDownStationId(), sectionRequest.getDistance());
+        sectionRepository.save(lineId, section.getUpStationId(), section.getDownStationId(), section.getDistance());
     }
 
     @Transactional
@@ -48,7 +48,7 @@ public class SectionService {
 
     // TODO : sections에게 현재 upId을 하행으로 가지고 있는 section이 있는지 물어봄
     // TODO : sections이 있다고 하면, 중간에 끼어 들어가는 상황
-        // TODO : sections로부터 Input upId를 출발지으로 가지고 있는 section의 도착지를 찾는다. (만약 가지고 있다면 중간삽입)
+    // TODO : sections로부터 Input upId를 출발지으로 가지고 있는 section의 도착지를 찾는다. (만약 가지고 있다면 중간삽입)
     private void addBaseOnUpStation(final Long lineId, final SectionRequest sectionRequest) {
         sectionRepository.saveBaseOnUpStation(lineId, sectionRequest);
     }
